Removed package "plugins" and introduced simpler module "plugin"
