+++ test/test_plugin.py
@@ -1,13 +1,22 @@
 import unittest
 
+from xcube.plugin import init_plugin
 from xcube.util.extension import ExtensionRegistry
 
 
-class CliCommandPluginTest(unittest.TestCase):
+class PluginTest(unittest.TestCase):
     def test_init_plugin(self):
-        from xcube.plugins.cli import init_plugin
         ext_reg = ExtensionRegistry()
+
         init_plugin(ext_reg)
+
+        self.assertTrue(ext_reg.has_extension('xcube.core.gen.iproc', 'default'))
+
+        self.assertTrue(ext_reg.has_extension('xcube.core.dsio', 'zarr'))
+        self.assertTrue(ext_reg.has_extension('xcube.core.dsio', 'netcdf4'))
+        self.assertTrue(ext_reg.has_extension('xcube.core.dsio', 'csv'))
+        self.assertTrue(ext_reg.has_extension('xcube.core.dsio', 'mem'))
+
         self.assertTrue(ext_reg.has_extension('xcube.cli', 'compute'))
         self.assertTrue(ext_reg.has_extension('xcube.cli', 'extract'))
         self.assertTrue(ext_reg.has_extension('xcube.cli', 'gen'))
