+++ test/util/test_plugin.py
@@ -38,9 +38,7 @@ class PluginTest(unittest.TestCase):
     def test_get_xcube_default_plugins(self):
         plugins = get_plugins()
         self.assertIsNotNone(plugins)
-        self.assertIn('xcube_default_iproc', plugins)
-        self.assertIn('xcube_default_dsio', plugins)
-        self.assertIn('xcube_default_cli', plugins)
+        self.assertIn('xcube_standard', plugins)
 
     def test_load_plugins_by_entry_points(self):
         plugins = load_plugins([EntryPoint('test', init_plugin)], ext_registry=self.ext_registry)
